Incremental updates
Add eye detection feature addressing false alarms. If only one
eye of the user is detected yet they are in bounds of the primary
attention zone, sound an alert. If the user is out of bounds but
both eyes are being detected, do NOT trigger an alert.

Video logging now supports accident handling. If an accident
occurs ("A") the two most recent log files are preserved and two
more are started.